+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod keys;
 mod message;
 mod options;
 pub mod pb;
+mod redis_broker;
 mod redis_plan;
 mod state;
 mod task;
@@ -15,6 +16,7 @@ pub use client::{
 pub use enqueue::{DEFAULT_MAX_RETRY, DEFAULT_TIMEOUT, EnqueuePlan, EnqueuePlanError};
 pub use message::{DecodeTaskMessageError, TaskMessage};
 pub use options::TaskOption;
+pub use redis_broker::{RedisBroker, RedisBrokerError, RedisExecutor, RedisExecutorError};
 pub use redis_plan::{
     RedisArg, RedisEnqueueOperation, RedisEnqueuePlan, RedisEnqueuePlanError, RedisEnqueueScript,
     RedisScriptCall,